xml-tool/backend/requirements.txt
Flask
flask-cors
psycopg2-binary
pandas
openpyxl
plyer
